Add files via upload
The Caesar cipher, also known as the Caesar shift or Caesar's code, is one of the simplest and most well-known encryption techniques. Named after Julius Caesar, who used it in his private correspondence, it is a type of substitution cipher where each letter in the plaintext is shifted a certain number of places down or up the alphabet.